+++ packages/react/src/components/exports.ts
@@ -1,3 +1,4 @@
 export * from "./Actions/exports"
 export * from "./Charts/exports"
+export * from "./layouts/exports"
 export * from "./Utilities/exports"
