+++ packages/react/src/components/layouts/StandardLayout/index.tsx
@@ -1,6 +1,6 @@
+import { cn } from "@/lib/utils"
 import { cva, type VariantProps } from "cva"
 import React from "react"
-import { cn } from "../../../lib/utils"
 import { LayoutProvider } from "../LayoutProvider"
 
 export interface StandardLayoutProps
