+++ packages/react/src/components/layouts/TwoColumnLayout/index.stories.tsx
@@ -1,11 +1,11 @@
 import type { Meta, StoryObj } from "@storybook/react"
 
+import { DetailsItemsList } from "@/experimental/Lists/DetailsItemsList"
+import * as DetailsItemsListStories from "@/experimental/Lists/DetailsItemsList/index.stories"
+import { Dashboard } from "@/experimental/Widgets/Layout/Dashboard"
+import * as DashboardStories from "@/experimental/Widgets/Layout/Dashboard/index.stories"
+import { PageDecorator } from "@/lib/storybook-utils/pageDecorator"
 import { ComponentProps } from "react"
-import { PageDecorator } from "../../../lib/storybook-utils/pageDecorator"
-import { Dashboard } from "../../Widgets/Layout/Dashboard"
-import * as DashboardStories from "../../Widgets/Layout/Dashboard/index.stories"
-import { DetailsItemsList } from "../Utils/DetailsItemsList"
-import * as DetailsItemsListStories from "../Utils/DetailsItemsList/index.stories"
 import { TwoColumnLayout } from "./index"
 
 const meta = {
