+++ packages/react/src/experimental/Lists/DetailsItem/index.tsx
@@ -1,7 +1,7 @@
 import { ComponentProps, FC, forwardRef } from "react"
-import { cn } from "../../../../lib/utils"
-import { DataList } from "../../../Lists/DataList"
-import { Weekdays } from "../../../Widgets/Content/Weekdays"
+import { cn } from "@/lib/utils"
+import { Weekdays } from "@/experimental/Widgets/Content/Weekdays"
+import { DataList } from "../DataList"
 
 type Content =
   | (ComponentProps<typeof DataList.Item> & {
