+++ packages/react/src/experimental/Lists/DetailsItemsList/index.tsx
