+++ packages/react/src/experimental/OneDataCollection/index.tsx
@@ -2,10 +2,10 @@ import { cn } from "@/lib/utils"
 import { motion } from "framer-motion"
 import { useEffect, useMemo, useState } from "react"
 import { useDebounceValue } from "usehooks-ts"
+import { useLayout } from "@/components/layouts/LayoutProvider"
 import { Icon } from "../../components/Utilities/Icon"
 import { Spinner } from "../../icons/app"
 import { OneActionBar } from "../OneActionBar"
-import { useLayout } from "../PageLayouts/LayoutProvider"
 import { CollectionActions } from "./CollectionActions/CollectionActions"
 import * as Filters from "./Filters"
 import type { FiltersDefinition, FiltersState } from "./Filters/types"
