--- packages/react/src/experimental/PageLayouts/Utils/exports.tsx
@@ -1,2 +1,0 @@
-export * from "./DetailsItem"
-export * from "./DetailsItemsList"
