+++ packages/react/src/experimental/exports.ts
@@ -11,7 +11,6 @@ export * from "./OneCard"
 export * from "./OneDataCollection/exports"
 export * from "./OnePagination"
 export * from "./Overlays/exports"
-export * from "./PageLayouts/exports"
 export * from "./RichText/exports"
 export * from "./Utilities/exports"
 export * from "./Widgets/exports"
